Move some functions into utils
