+++ runners/kiosk.py
@@ -1,7 +1,6 @@
 import random
 import sys
 import time
-from importlib import import_module
 from queue import Empty
 
 from loguru import logger
@@ -9,42 +8,6 @@ from loguru import logger
 import broadcasters
 import controllers
 from runners import utils
-
-
-def load_demo(module_name):
-    """
-    Given a module name it will load the module and get the modules demo
-    class.
-
-    Args:
-        module_name (str): Name of the module to load.
-
-    Returns:
-        class: The demo class.
-    """
-    logger.debug(f"Loading {module_name}")
-    return utils.get_demo_cls(import_module(module_name))
-
-
-def load_demos(demo_dir="demos"):
-    """
-    Loads all demos for a given directory. It returns the demos in a dictionary
-    with the name of the demo as key and the demo module as the value.
-
-    Args:
-        demo_dir (str): Directory where the demos are located.
-
-    Returns:
-        dict: Dictionary with the name of the demo as key and the demo module
-    """
-    logger.debug("Loading demos...")
-
-    demos = utils.get_demos(demo_dir)
-
-    # Load the module
-    demos = {name: load_demo(module) for name, module in demos}
-
-    return demos
 
 
 def get_random_demo(demos):
@@ -202,7 +165,7 @@ def run_loop(screen, user_input_timeout=300, demo_time_override=None):
 
     queues = utils.Queues()
 
-    demos = load_demos()
+    demos = utils.load_demos()
     random_demos = get_random_demo(demos)
     handle_input = controllers.start_inputs(
         queues.system_queue, queues.demo_input_queue
