+++ runners/utils.py
@@ -1,6 +1,9 @@
 from dataclasses import dataclass
+from importlib import import_module
 from pathlib import Path
 from queue import Queue
+
+from loguru import logger
 
 
 @dataclass
@@ -53,3 +56,39 @@ def get_demo_cls(demo_module):
     return getattr(
         demo_module, "".join([word.capitalize() for word in demo_name.split("_")])
     )
+
+
+def load_demo(module_name):
+    """
+    Given a module name it will load the module and get the modules demo
+    class.
+
+    Args:
+        module_name (str): Name of the module to load.
+
+    Returns:
+        class: The demo class.
+    """
+    logger.debug(f"Loading {module_name}")
+    return get_demo_cls(import_module(module_name))
+
+
+def load_demos(demo_dir="demos"):
+    """
+    Loads all demos for a given directory. It returns the demos in a dictionary
+    with the name of the demo as key and the demo module as the value.
+
+    Args:
+        demo_dir (str): Directory where the demos are located.
+
+    Returns:
+        dict: Dictionary with the name of the demo as key and the demo module
+    """
+    logger.debug("Loading demos...")
+
+    demos = get_demos(demo_dir)
+
+    # Load the module
+    demos = {name: load_demo(module) for name, module in demos}
+
+    return demos
